moved loading code into new module
move target list to config

relint

return frames
