+++ src/seismometer/seismogram.py
@@ -1,18 +1,17 @@
 import json
 import logging
-import warnings
 from functools import lru_cache
 from pathlib import Path
 from typing import Optional
 
 import numpy as np
 import pandas as pd
-import pyarrow.parquet as pq
 
 from seismometer.configuration import ConfigProvider
 from seismometer.core.patterns import Singleton
 from seismometer.data import pandas_helpers as pdh
 from seismometer.data import resolve_cohorts
+from seismometer.data.loader import seismogram_loader_factory
 from seismometer.report.alerting import AlertConfigProvider
 
 MAXIMUM_NUM_COHORTS = 25
@@ -72,12 +71,20 @@ class Seismogram(object, metaclass=Singleton):
             config_path = Path(config_path)
 
         self.cohort_cols: list[str] = []
-        self.target_cols: list[str] = []
         self.config_path = config_path
-        self._load_from_config(config_path)
+
+        self.load_config(config_path)
 
         self.config.set_output(output_path)
         self.config.output_dir.mkdir(parents=True, exist_ok=True)
+        self.dataloader = seismogram_loader_factory(self.config)
+
+    def load(self, predictions=None, events=None):
+        self._load_metadata()
+        self.dataframe = self.dataloader.load(predictions, events)
+
+        self.create_cohorts()
+        self._set_df_counts()  # cache all stat counts before we clean up memory
 
         # UI Controls
         if self.cohort_cols:
@@ -139,6 +146,10 @@ class Seismogram(object, metaclass=Singleton):
         if event.endswith("_Value"):
             event = event[:-6]
         self._target = event
+
+    @property
+    def target_cols(self) -> list:
+        return self.config.target_cols
 
     @property
     def intervention(self):
@@ -267,8 +278,7 @@ class Seismogram(object, metaclass=Singleton):
     # endregion
 
     # region initialization and preprocessing (this region knows about config)
-
-    def _load_from_config(self, config_path: str | Path):
+    def load_config(self, config_path: Path):
         self.config = ConfigProvider(config_path)
         self.alert_config = AlertConfigProvider(config_path)
 
@@ -280,9 +290,7 @@ class Seismogram(object, metaclass=Singleton):
         self.predict_time = self.config.predict_time
         self.output_list = self.config.output_list
         self.target_event = self.config.target
-
-        self._load_metadata()
-        self._load_data()
+        self._cohorts = self.config.cohorts
 
     def _load_metadata(self):
         """
@@ -298,134 +306,6 @@ class Seismogram(object, metaclass=Singleton):
             self.thresholds = [0.8, 0.5]
 
         self.modelname: str = self._metadata.get("modelname", "UNDEFINED MODEL")
-
-    def _load_predictions(self):
-        """
-        Loads the predictions data, restricting features based on config (if any).
-        """
-        if self.config.features:  # no features == all features
-            desired_columns = set(self.config.prediction_columns)
-
-            with warnings.catch_warnings():
-                warnings.simplefilter("ignore", category=FutureWarning)
-                present_columns = set(
-                    pq.ParquetDataset(self.config.prediction_path, use_legacy_dataset=False).schema.names
-                )
-
-            if self.config.target in present_columns:
-                desired_columns.add(self.config.target)
-
-            actual_columns = desired_columns & present_columns
-            if len(desired_columns) != len(actual_columns):
-                logger.warning(
-                    "Not all requested columns are present. "
-                    + f"Missing columns are {', '.join(desired_columns-present_columns)}"
-                )
-                logger.debug(f"Requested columns are {', '.join(desired_columns)}")
-                logger.debug(f"Columns present are {', '.join(present_columns)}")
-            self.dataframe = pd.read_parquet(self.config.prediction_path, columns=actual_columns)
-        else:
-            self.dataframe = pd.read_parquet(self.config.prediction_path)
-
-        if self.config.target in self.dataframe:
-            logger.debug(
-                f"Using existing column in predictions dataframe as target: {self.config.target} -> {self.target}"
-            )
-            self.dataframe = self.dataframe.rename({self.config.target: self.target}, axis=1)
-
-    def _load_events(self):
-        """
-        Loads the events data if any exists, otherwise stands up an empty df with the expected columns.
-        """
-        try:
-            self._events = pd.read_parquet(self.config.event_path).rename(
-                columns={self.config.ev_type: "Type", self.config.ev_time: "Time", self.config.ev_value: "Value"},
-                copy=False,
-            )
-        except BaseException:
-            logger.debug(f"No events found at {self.config.event_path}")
-            self._events = pd.DataFrame(columns=self.entity_keys + ["Type", "Time", "Value"])
-
-    def _load_data(self):
-        """
-        Loads data and does some preprocessing.
-        """
-        # Probably don't want to always read these into memory; reloadable
-        logger.info(f"Importing files from {self.config.config_dir}")
-
-        self._load_predictions()
-        self._load_events()
-
-        self._time_to_ns()
-
-        self._cohorts = self.config.cohorts
-        self.prep_data()
-
-    def _time_to_ns(self):
-        # datetime precisions don't play nicely - fix to pands default
-        pred_times = self.dataframe.select_dtypes(include="datetime").columns
-        self.dataframe[pred_times] = self.dataframe[pred_times].astype({col: "<M8[ns]" for col in pred_times})
-
-        # Time column in events is known
-        self._events["Time"] = self._events["Time"].astype("<M8[ns]")
-
-    def prep_data(self):
-        """Preprocesses the prediction data for use in analysis."""
-        # Do not infer events
-        self.dataframe = self._infer_datetime(self.dataframe)
-
-        # Expand this to robust score prep
-        for score in self.output_list:
-            if score not in self.dataframe:
-                continue
-            if 50 < self.dataframe[score].max() <= 100:  # Assume out of 100, readjust
-                self.dataframe[score] /= 100
-
-        # Need to remove pd.FloatXxDtype as sklearn and numpy get confused
-        float_cols = self.dataframe.select_dtypes(include=[float]).columns
-        self.dataframe[float_cols] = self.dataframe[float_cols].astype(np.float32)
-
-        self.add_events()
-        self.create_cohorts()
-        self._set_df_counts()  # cache all stat counts before we clean up memory
-        del self._events  # clean up events dataframe since all events have been merged
-
-    def add_events(self) -> None:
-        """Merges a value and time column into dataframe for each configured event."""
-        # -> should specify a specific outcome
-        self.dataframe = (
-            self.dataframe.sort_values(self.predict_time)
-            .drop_duplicates(subset=self.entity_keys + [self.predict_time])
-            .dropna(subset=[self.predict_time])
-        )
-        self.events = self.events.sort_values("Time")
-
-        for event in self.config.events:
-            # Merge
-            if event.window_hr:
-                logger.debug(
-                    f"Windowing event {event.display_name} to lookback {event.window_hr} offset by {event.offset_hr}"
-                )
-                self.dataframe = self._merge_event(
-                    event.source,
-                    frozenset(self.dataframe.columns),
-                    window_hrs=event.window_hr,
-                    offset_hrs=event.offset_hr,
-                    display=event.display_name,
-                    sort=False,
-                )
-                self.target_cols.append(event.display_name)
-            else:  # No lookback
-                logger.debug(f"Merging event {event.display_name}")
-                self.dataframe = self._merge_event(
-                    event.source, frozenset(self.dataframe.columns), display=event.display_name, sort=False
-                )
-
-            # Impute
-            if event.impute_val or event.usage == "target":
-                event_val = pdh.event_value(event.display_name)
-                impute = event.impute_val or 0  # Enforce binary type target
-                self.dataframe[event_val] = self.dataframe[event_val].fillna(impute)
 
     def create_cohorts(self) -> None:
         """Creates data columns for each cohort defined in configuration."""
@@ -461,22 +341,6 @@ class Seismogram(object, metaclass=Singleton):
             self.cohort_cols.append(disp_attr)
         logger.debug(f"Created cohorts: {', '.join(self.cohort_cols)}")
 
-    def _merge_event(self, event, frame_columns, offset_hrs=0, window_hrs=None, display="", sort=True):
-        disp_event = display if display else event
-        translate_event = event if display else ""
-
-        return pdh.merge_windowed_event(
-            self.dataframe,
-            self.predict_time,
-            self.events.replace({"Type": translate_event}, disp_event),
-            disp_event,
-            self.entity_keys,
-            min_leadtime_hrs=offset_hrs,
-            window_hrs=window_hrs,
-            event_base_time_col="Time",
-            sort=sort,
-        )
-
     @lru_cache
     def _data_mask(self, event_val):
         return self.dataframe[event_val] != -1
@@ -486,16 +350,5 @@ class Seismogram(object, metaclass=Singleton):
         # Require events with time or negative label
         neg_mask = np.ones(keep_zero).astype(bool) if keep_zero is None else self.dataframe[keep_zero] == 0
         return self.dataframe[event_time].notna() | neg_mask
-
-    @staticmethod
-    def _infer_datetime(df, cols=None, override_categories=None):
-        # override_categories - allow configured dtypes to force decision
-        if cols is None:
-            cols = df.columns
-        for col in cols:
-            if "Time" in col:
-                df[col] = pd.to_datetime(df[col])
-                continue
-        return df
 
     # endregion
